backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
qdrant-client>=1.9.0
cohere>=5.5.0
anthropic>=0.28.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
boto3>=1.34.0
python-docx>=1.1.0
reportlab>=4.2.0
unstructured[all-docs]>=0.14.0
tiktoken>=0.7.0
slowapi>=0.1.9
structlog>=24.1.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
aiofiles>=23.2.0
email-validator>=2.1.0
